增加ToMaven任务调试完成

+++ Gradle/src/main/java/com/pqixing/modularization/utils/ResultUtils.kt
@@ -4,6 +4,7 @@ import com.pqixing.Tools
 import com.pqixing.modularization.Keys
 import com.pqixing.modularization.manager.ManagerPlugin
 import com.pqixing.tools.FileUtils
+import com.pqixing.tools.TextUtils
 import org.gradle.api.GradleException
 import java.io.File
 
@@ -12,26 +13,29 @@ import java.io.File
  */
 object ResultUtils {
 
+    //是否是通过ide调用
+    val ide = "ide" == TextUtils.getSystemEnv("syncType")
     /**
      * 输出结果，用于Ide的交互获取
      * @param exitCode 退出标记 0 表示正常退出，1表示异常退出
      */
-    fun writeResult(msg: String, exitCode: Int = 0) {
-        val plugin = ManagerPlugin.getPlugin()
-        val ideFile = File(plugin.cacheDir, "ide.log")
-        val log = "${Keys.PREFIX_IDE_LOG}?${Keys.RUN_TASK_ID}=${getProperty(Keys.RUN_TASK_ID)
-                ?: System.currentTimeMillis()}&endTime=${System.currentTimeMillis()}&exitCode=$exitCode&msg=$msg"
-
+    fun writeResult(msg: String, exitCode: Int = 0,exit:Boolean = exitCode!=0) {
         Tools.println(msg)
-        if (ideFile.exists() && ideFile.length() > 102400) {
-            Tools.println("writeResult del ${ideFile.path}-> ${ideFile.length()}")
-            FileUtils.delete(ideFile)
-        }
+        if(ide) {
+            val plugin = ManagerPlugin.getPlugin()
+            val ideFile = File(plugin.rootDir, ".idea/modularization.log")
+            val log = "${Keys.PREFIX_IDE_LOG}?${Keys.RUN_TASK_ID}=${getProperty(Keys.RUN_TASK_ID)
+                    ?: System.currentTimeMillis()}&endTime=${System.currentTimeMillis()}&exitCode=$exitCode&msg=$msg"
+            if (ideFile.exists() && ideFile.length() > 102400) {
+                Tools.println("writeResult del ${ideFile.path}-> ${ideFile.length()}")
+                FileUtils.delete(ideFile)
+            }
 
-        if (!ideFile.exists()) {
-            FileUtils.writeText(ideFile, log)
-        } else ideFile.appendText("\n" + log)
-        if (exitCode != 0) {
+            if (!ideFile.exists()) {
+                FileUtils.writeText(ideFile, log)
+            } else ideFile.appendText("\n" + log)
+        }
+        if (exit) {
             throw  GradleException(msg)
         }
     }
